# perf: optimize getGenericTemplate in GenericTemplateDetails