--- a/apps/frontend/src/graphql/genericTemplates/getGenericTemplateMinimal.graphql
+++ b/apps/frontend/src/graphql/genericTemplates/getGenericTemplateMinimal.graphql
@@ -1,0 +1,6 @@
+query getGenericTemplateMinimal($genericTemplateId: Int!) {
+  genericTemplate(genericTemplateId: $genericTemplateId) {
+    title
+    questionaryId
+  }
+}