--- a/apps/frontend/src/hooks/genericTemplate/useGenericTemplate.ts
+++ b/apps/frontend/src/hooks/genericTemplate/useGenericTemplate.ts
@@ -5,14 +5,16 @@ import { useDataApi } from 'hooks/common/useDataApi';
 import { GenericTemplateFragment } from '../../generated/sdk';
 
 export function useGenericTemplate(genericTemplateId: number) {
-  const [genericTemplate, setGenericTemplate] =
-    useState<GenericTemplateFragment | null>(null);
+  const [genericTemplate, setGenericTemplate] = useState<Pick<
+    GenericTemplateFragment,
+    'title' | 'questionaryId'
+  > | null>(null);
 
   const api = useDataApi();
 
   useEffect(() => {
     api()
-      .getGenericTemplate({ genericTemplateId })
+      .getGenericTemplateMinimal({ genericTemplateId })
       .then((data) => {
         if (data.genericTemplate) {
           setGenericTemplate(data.genericTemplate);