Closes UserOfficeProject/issue-tracker#1257

## Description
This PR optimizes the `getGenericTemplate` function in `GenericTemplateDetails`.

2 queries against prod

| getGenericTemplate  | getGenericTemplateMinimal |
| ------------- | ------------- |
| 0.195345s | 0.134789s |
| 0.178732s | 0.127624s |
| 0.196071s | 0.132472s |
| 0.201398s | 0.126692s |
| 0.181837s | 0.115354s |
| 0.224866s | 0.110462s |
| 0.192504s | 0.110154s |
| 0.210028s | 0.136687s |
| 0.181648s | 0.114990s |
| 0.200214s | 0.175416s |

## Motivation and Context
The existing `getGenericTemplate` function was fetching more data than necessary, causing performance issues. This change is required to improve the efficiency of the function and overall performance of the application.

## Changes
- Introduced a new query `getGenericTemplateMinimal` that fetches only the `title` and `questionaryId` fields, instead of the entire `GenericTemplateFragment`.
- Updated the `useGenericTemplate` hook to use the new `getGenericTemplateMinimal` query, reducing the amount of data fetched and processed.
- Updated the state initialization in `useGenericTemplate` to reflect the reduced data structure.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Fixes Jira Issue

## Depends On

<!--- Does this PR depend on another PR that should be merged first or at the same time -->

## Tests included/Docs Updated?

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->

- [ ] I have added tests to cover my changes.
- [ ] All relevant doc has been updated